fix: resolve golangci-lint code quality issues
- Fix comment formatting (add space after //)
- Remove unnecessary int32 conversion
- Replace if-else chain with switch statement

These changes address the golangci-lint errors that were causing
the CI pipeline to fail.

+++ internal/server/asset.go
@@ -58,7 +58,7 @@ func (s *Server) GetAssets(ctx context.Context, request *immichv1.GetAssetsReque
 
 	assets, err := s.db.GetAssets(ctx, sqlc.GetAssetsParams{
 		OwnerId:    userID,
-		Limit:      int32(request.Size),
+		Limit:      request.Size,
 		Offset:     offset,
 		Type:       assetType,
 		IsFavorite: isFavorite,
